Add 3.5 to the list of supported versions
Currently, `setup.py` errors out on python 3.5, because it refuses to install on any version of Python that isn't explicitly listed in the Trove classifiers. This adds 3.5 to the list of supported versions.

This appears to be the only change required to support 3.5 -- after making this change to a 1.2.4 source tree then I tested on Linux and was able to build a wheel on 3.5 that built without errors and seems to be working fine in cursory testing (e.g. starting up matplotlib and opening a plot window, using PySide as the gui backend).